% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{basic_axis}
\alias{basic_axis}
\title{Draw a Custom Axis with Formatted Labels}
\usage{
basic_axis(
  format = "integer",
  sides = "bottom",
  line_width = 0.5,
  ...,
  at = NULL
)
}
\arguments{
\item{format}{A character string specifying the format of the axis labels. Options include:
`"percent"`, `"dollar"`, `"date"`, `"POSIXt"`, `"integer"`, and `"scientific"`. Default is `"integer"`.}

\item{sides}{A character vector indicating which sides of the plot to draw the axis on.
Options are `"bottom"`, `"left"`, `"top"`, and `"right"`. Default is `"bottom"`.}

\item{line_width}{Numeric. The width of the axis line. Default is `0.5`.}

\item{...}{Additional graphical parameters passed to the `axis()` function.}

\item{at}{Optional. A numeric vector specifying the exact positions for ticks. If `NULL`,
positions are determined automatically using `pretty()` on `axTicks()`.}
}
\value{
The function is used for its side effects (modifying the plot). It does not
return a value.
}
\description{
This function adds a custom axis to a base graphics plot with user-defined formatting
for tick labels. It supports multiple formats, such as percentages, currency, dates,
and scientific notation.
}
\details{
The function creates a custom axis with tick marks on the specified sides of the plot.
The labels can be formatted according to different styles using the `format` parameter.

The axis labels are formatted using the `format_axis()` function, allowing for different
display formats, including percentages, currency, and dates.
}
\examples{
# Create a basic plot without default axes
plot(mpg ~ wt, data = mtcars, type = "p", axes = FALSE)
basic_axis(format = "integer", sides = "bottom")
basic_axis(format = "percent", sides = "left")

}
\seealso{
\code{\link{axis}}, \code{\link{format_axis}}
}
