% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{basic_palettes}
\alias{basic_palettes}
\title{List of Cuban paintings color palettes and the order in which they're printed}
\format{
An object of class \code{list} of length 12.
}
\usage{
basic_palettes
}
\description{
Complete list of Cuban paintings color palette
}
\details{
Use names(basic_palettes) to return a list of all possible palette names. Currently,
the following palettes are available:
\code{imagenes},\code{primvera_descanso},\code{no_aguanto_mas},\code{emilio_rodriguez},
\code{paisaje_bohio},\code{productivismo}, \code{retrato_marin}, \code{helena_herrera},
\code{emarque_colon}, \code{havana}, \code{guajiros} and \code{el_humo}.
Use \code{\link{basic_palettes}} to construct palettes.
}
\keyword{datasets}
