% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Edward Tufte Style Boxplot}
\usage{
box_plot(
  formula,
  .data,
  labels = TRUE,
  label_names = NULL,
  label_size = 1.2,
  main = NULL,
  in_line = 1,
  xlab = NULL,
  ylab = NULL,
  n_ticks = 5,
  horizontal = TRUE,
  line_type = "solid",
  line_color = "gray20",
  median_symbol = "|",
  median_color = "black",
  add_mean = TRUE,
  mean_symbol = 1,
  mean_color = "red",
  outlier_symbol = 1,
  outlier_size = 0.5,
  ...
)
}
\arguments{
\item{formula}{A formula of the form `y ~ x`, where `y` is the grouping variable, and `x` is the numeric variable(s).}

\item{.data}{A data frame containing the variables specified in the formula.}

\item{labels}{Logical; if `TRUE`, adds axis labels corresponding to the grouping variable.}

\item{label_names}{Optional vector of names for the labels (defaults to the levels of `y`).}

\item{label_size}{Numeric; size of the axis labels.}

\item{main}{Character; title of the plot.}

\item{in_line}{Numeric; line spacing for the title.}

\item{xlab}{Character; labels for the x- and y-axes.}

\item{ylab}{Character; labels for the x- and y-axes.}

\item{n_ticks}{Numeric; number of ticks for the axis representing the numeric variable.}

\item{horizontal}{Logical; if `TRUE`, creates a horizontal boxplot (default is `TRUE`).}

\item{line_type}{Character; type of line for whiskers and box edges (e.g., `"solid"`).}

\item{line_color}{Character; color of the boxplot lines (default is `"gray20"`).}

\item{median_symbol}{Character or numeric; symbol used to represent the median.}

\item{median_color}{Character; color of the median symbol.}

\item{add_mean}{Logical; if `TRUE`, adds a marker for the mean value.}

\item{mean_symbol}{Character or numeric; symbol used to represent the mean (default is `1`).}

\item{mean_color}{Character; color of the mean symbol (default is `"red"`).}

\item{outlier_symbol}{Character or numeric; symbol used to represent outliers.}

\item{outlier_size}{Numeric; size of the outlier symbols.}

\item{...}{Additional graphical parameters passed to `par`.}
}
\value{
A Tufte-style boxplot is drawn on the active graphical device.
}
\description{
Draws a minimalist boxplot inspired by Edward Tufte's visualization principles.
The function supports customization for horizontal or vertical orientation,
additional mean markers, and labeled groups.
}
\details{
This function generates a boxplot with reduced visual clutter, suitable for data comparison in publications or presentations.
Outliers are displayed as individual points, and optional mean markers can be included.
}
\examples{
# Basic Example
data(mtcars)
box_plot(cyl~ mpg, .data = mtcars, main = "MPG by Cylinder")

# Horizontal Example with Mean Marker
box_plot(Diet~weight, .data = ChickWeight,
horizontal = FALSE, median_symbol = "×",
add_mean = TRUE, n_ticks = 7,
ylab = "Diet",
xlab = "Weight(g)")

}
