% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_histogram.R
\name{cut_counts}
\alias{cut_counts}
\title{Count Observations in Each Bin}
\usage{
cut_counts(x, breaks, right = TRUE, include.lowest = TRUE)
}
\arguments{
\item{x}{A numeric vector representing the data to be binned.}

\item{breaks}{A numeric vector of bin edges, or a scalar for the number of bins.}

\item{right}{Logical, whether bins should be right-closed. Default is `TRUE`.}

\item{include.lowest}{Logical, whether the lowest bin should include the smallest value. Default is `TRUE`.}
}
\value{
An integer vector containing the count of observations in each bin.
}
\description{
Counts the number of observations within each bin of a histogram, given a vector of bin edges (`breaks`).
}
\details{
The function ensures that the `breaks` vector is strictly increasing. It uses the `cut()` function to assign bins to observations.
}
\examples{
x <- rnorm(100)
breaks <- seq(-3, 3, by = 1)
cut_counts(x, breaks)

}
