% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_histogram.R
\name{histogram}
\alias{histogram}
\title{Edward Tufte Style Histogram}
\usage{
histogram(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the data to be visualized.}

\item{...}{Additional arguments passed to the function.}

\item{shape}{The shape of the histogram: "segment", "rect", or "lines".}

\item{show_border}{A character string specifying which borders to display for shape = "segment" ("left" or "right").}

\item{na_rm}{Logical, whether to remove `NA` values from the data. Default is `TRUE`.}

\item{breaks}{Method to compute the histogram bins. Options are:
- `"sturges"`: Sturges' rule.
- `"scott"`: Scott's normal reference rule.
- `"fd"`: Freedman-Diaconis rule.
- `"doane"`: Doane's formula.
- `"ts"`: Terrell-Scott rule.
- A numeric value for fixed bin width or a custom function.}

\item{freq}{Logical, whether to display frequency (`TRUE`) or density (`FALSE`) on the y-axis. Default is `TRUE`.}

\item{right}{Logical, whether bins should be right-closed. Default is `TRUE`.}

\item{fuzz}{Numeric, a small value added to bin boundaries to prevent rounding issues.}

\item{include_lowest}{Logical, whether the lowest bin should include the smallest value. Default is `TRUE`.}

\item{axes}{Logical, whether to display x and y axes. Default is `FALSE`.}

\item{labels}{Logical, whether to display axis labels. Default is `TRUE`.}

\item{label_names}{Optional, custom names for x and y axis labels. Default is `NULL`.}

\item{label_size}{Numeric, size of axis labels. Default is `NULL`.}

\item{main}{Optional, a title for the histogram. Default is `NULL`.}

\item{in_line}{Numeric, line spacing for the title. Default is `1`.}

\item{xlab}{Custom x-axis label. Default is `NULL`.}

\item{ylab}{Custom y-axis label. Default is `NULL`.}

\item{line_width}{Numeric, line width for the histogram. Default is `0.5`.}

\item{line_type}{Line type for the histogram ("solid", "dashed", etc.). Default is `"solid"`.}

\item{line_color}{Color of the histogram lines. Default is `"gray20"`.}
}
\value{
A histogram plot with the desired styling and a list structure containing:
- `breaks`: The computed breaks for the histogram bins.
- `counts`: The frequency of observations in each bin.
- `density`: The density of observations in each bin.
- `mid_points`: The midpoints of each bin.
}
\description{
Visualize the distribution of a single **continuous** variable using Edward Tufte's minimalist principles for data visualization.

The function includes various methods for calculating histogram breaks, including:
- **Sturges**: Sturges' formula based on a normal distribution approximation.
- **Scott**: Scott's rule, which minimizes the integrated mean squared error for a normal density.
- **FD**: The Freedman-Diaconis rule, which uses the interquartile range for robust bin width calculation.
- **Doane**: A modified Sturges' formula to account for skewness in the data.
- **Terrell-Scott**: Terrell-Scott's non-normal reference rule for optimal binning, based on IQR.
}
\details{
The choice of breaks method significantly influences the histogram's appearance. The function provides robust defaults while allowing customization to suit specific data distributions.
}
\examples{
set.seed(123)
x <- rnorm(100)
histogram(x, shape = "rect", breaks = "fd", axes = TRUE, labels = TRUE)

# Using formula argument
df <- data.frame(x = rexp(1000,2.5))
histogram(~x,df,shape = "line", breaks = 12)

}
\seealso{
[hist()], [cut()],[geom_histogram()]
}
