% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{range_frame_axis}
\alias{range_frame_axis}
\title{Range Frame Axis with Summary Statistics and Custom Ticks}
\usage{
range_frame_axis(
  .x,
  sides = "bottom",
  at = NULL,
  draw_lines = FALSE,
  line_width = 0,
  tick_width = 0.7,
  tick_gap = 0.5,
  tick_length = -0.5,
  tick_color = "gray30",
  mean_color = "red",
  axis_gap = 0.05,
  median_gap = 0.05,
  digits = 2,
  label_size = 0.9,
  ...
)
}
\arguments{
\item{.x}{A numeric vector containing the data for which the summary statistics are calculated.}

\item{sides}{A character vector specifying which sides of the plot to draw the axis on.
Options include `"bottom"`, `"left"`, `"top"`, and `"right"`. Defaults to `c("bottom", "left")`.}

\item{at}{Optional numeric vector specifying tick positions. If `NULL`,
ticks are generated based on the summary statistics of `.x`.}

\item{draw_lines}{Logical indicating whether to draw lines extending from the axis for
the lower and upper quartiles. Defaults to `FALSE`.}

\item{line_width}{Numeric value specifying the width of the drawn lines if `draw_lines = TRUE`.
Defaults to `0`.}

\item{tick_width}{Numeric value for the width of tick marks. Defaults to `0.7`.}

\item{tick_gap}{Numeric value specifying the minimum gap between tick marks, in plot coordinates.
Defaults to `0.5`.}

\item{tick_length}{Numeric value specifying the length of tick marks. Negative values indicate
that ticks are drawn inward from the axis. Defaults to `-0.5`.}

\item{tick_color}{Color for the tick marks. Defaults to `"gray30"`.}

\item{mean_color}{Color for the label of the mean value. Defaults to `"red"`.}

\item{axis_gap}{Numeric value specifying the gap between the axis and the Q1/Q3 lines.
Used for both linear and logarithmic scales. Defaults to `0.05`.}

\item{median_gap}{Numeric value specifying the gap between the median and its label.
Used for both linear and logarithmic scales. Defaults to `0.05`.}

\item{digits}{Integer specifying the number of decimal places to round the tick labels.
Defaults to `2`.
@param label_size size of axis labels to be passed to `mtext`. Defaults to `0.9`}

\item{...}{Additional arguments to be passed to other plotting functions.}
}
\value{
Invisibly returns `NULL`, as the primary purpose is to modify the existing plot.
}
\description{
Adds a minimalist axis to a plot, inspired by Edward Tufte's range frames,
displaying summary statistics (min, Q1, median, mean, Q3, max) and custom ticks.
Supports both linear and logarithmic scales, with customizable colors,
tick positions, and axis properties.
}
\details{
This function is designed for minimalist data visualization, focusing on the
essential summary statistics of the data. It is especially useful for emphasizing
key statistical values while maintaining a clean, uncluttered plot appearance.

The function handles both linear and logarithmic scales seamlessly, automatically
adjusting the positioning of summary statistics, labels, and tick marks based on
the scale of the axis.

If the `draw_lines` parameter is set to `TRUE`, the function draws lines extending
from the axis to represent the quartiles. The axis labels include the mean, median,
Q1, and Q3 values, with the mean highlighted using the specified `mean_color`.
}
\seealso{
\code{\link{axis}}, \code{\link{summary}}, \code{\link{lines}}
}
