% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_font.R
\name{reset_font}
\alias{reset_font}
\title{Reset previous or default font used in every graphing device}
\usage{
reset_font(family)
}
\arguments{
\item{family}{Optional argument to reset fonts to previously used font. No argument needed
if setting font to default system fonts}
}
\value{
message confirming default font change
}
\description{
Reset fonts to default system fonts, 'sans','serif', or 'mono' or previously used font
}
\examples{
set_plot_font("Libre Franklin")

hist(
rnorm(100),
xlab = "x-axis",
ylab = "y-axis"
)

reset_plot_font()

hist(
rnorm(100),
xlab = "x-axis",
ylab = "y-axis"
)
}
