% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_histogram.R
\name{scott_breaks}
\alias{scott_breaks}
\title{Scott's Normal Reference Rule for Number of Bins}
\usage{
scott_breaks(x)
}
\arguments{
\item{x}{A numeric vector representing the data to be binned.}
}
\value{
An integer representing the number of bins.
}
\description{
Computes the number of histogram bins using Scott's normal reference rule, which minimizes the integrated mean squared error for a normal density.
}
\examples{
x <- rnorm(100)
scott_breaks(x)

}
