% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_font.R
\name{set_font}
\alias{set_font}
\title{Set Google font for base R and ggplot2 plots}
\usage{
set_font(name, family)
}
\arguments{
\item{name}{Google font name}

\item{family}{font family name to be used in R. 'family' can take any character argument.
If no family name provided, R will use the Google font name as default.}
}
\value{
temporarily invisible copy of Google font name
}
\description{
Set Google font as default font for all graphing devices
}
\examples{
set_font("Montserrat","monse")

hist(
rnorm(100),
xlab = "x-axis"),
ylab = "y-axis"
)
}
