% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\title{Set custom theme for base R plots}
\usage{
set_theme(...)
}
\arguments{
\item{...}{arguments to be passed as \code{parameter = value} pairs}
}
\value{
a list of all theme settings, invisibly.
}
\description{
Sets and returns base plotting theme parameters
}
\details{
1. The user can pass no arguments, in which case, the current theme will be returned
as a list \cr
2. NULL - the current theme settings will be removed \cr
3. A list can be passed as an argument. The new list will supersede the current theme.
}
\examples{
# Set theme by name
set_theme("crisp")
plot(0:10,0:10)

# Get list of theme parameters
set_theme()

# Reset theme
set_theme(NULL)

}
