% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_histogram.R
\name{sturges_breaks}
\alias{sturges_breaks}
\title{Sturges Number-of-Bins Rule}
\usage{
sturges_breaks(x)
}
\arguments{
\item{x}{A numeric vector representing the data to be binned.}
}
\value{
An integer representing the number of bins.
}
\description{
Computes the number of histogram bins using Sturges' formula, which is based on a normal distribution approximation.
}
\examples{
x <- rnorm(100)
sturges_breaks(x)

}
