% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_histogram.R
\name{ts_breaks}
\alias{ts_breaks}
\title{Terrell-Scott Oversmoothed Histogram Formula}
\usage{
ts_breaks(x, use_iqr = FALSE)
}
\arguments{
\item{x}{A numeric vector representing the data to be binned.}

\item{use_iqr}{Logical, whether to use the interquartile range (IQR) for a robust version of the formula. Default is `FALSE`.}
}
\value{
An integer representing the number of bins.
}
\description{
Computes the number of histogram bins using Terrell-Scott's oversmoothing method, which provides a larger bin width to reveal more structure in the data.
}
\examples{
x <- rnorm(100)
ts_breaks(x, use_iqr = TRUE)

}
